+++ contracts/Test-helpers/TestERC20.sol
@@ -19,12 +19,12 @@ contract TestERC20 is ITestERC20, ERC20 {
         permitted[user] = permit;
     }
 
-    function mint(address account, uint amount) external override {
+    function mint(address account, uint256 amount) external override {
         require(permitted[msg.sender], "only permitted");
         ERC20._mint(account, amount);
     }
 
-    function burn(address account, uint amount) external override {
+    function burn(address account, uint256 amount) external override {
         require(permitted[msg.sender], "only permitted");
         ERC20._burn(account, amount);
     }
