+++ contracts/periphery/MetaDefenderMarketsRegistry.sol
@@ -1,39 +1,44 @@
 //SPDX-License-Identifier: ISC
 pragma solidity 0.8.9;
 
-import "@openzeppelin/contracts/access/Ownable.sol";
-import "@openzeppelin/contracts/utils/structs/EnumerableSet.sol";
-import "../interfaces/IMetaDefenderMarketsRegistry.sol";
+import '@openzeppelin/contracts/access/Ownable.sol';
+import '@openzeppelin/contracts/utils/structs/EnumerableSet.sol';
+import '../interfaces/IMetaDefenderMarketsRegistry.sol';
 
 /**
  * @title MetaDefenderMarketsRegistry
  * @author MetaDefender
  * @dev Registry that allow external services to keep track of the deployments MetaDefender Markets
  */
-contract MetaDefenderMarketsRegistry is Ownable, IMetaDefenderMarketsRegistry{
+contract MetaDefenderMarketsRegistry is Ownable, IMetaDefenderMarketsRegistry {
     using EnumerableSet for EnumerableSet.AddressSet;
 
     EnumerableSet.AddressSet internal insuranceMarkets;
     mapping(address => MarketAddresses) public insuranceMarketsAddresses;
 
-    event MarketAdded(address metaDefender, address liquidityCertificate, address policy, address epochManage);
+    event MarketAdded(
+        address metaDefender,
+        address liquidityCertificate,
+        address policy,
+        address epochManage
+    );
     event MarketRemoved(address metaDefender);
 
     /**
      * @dev Method to register the addresses of a new deployments market
-   *
-   * @param metaDefender Address of the metaDefender contract
-   * @param liquidityCertificate Address of the liquidityCertificate contract(ERC721)
-   * @param policy Address of the policy contract(ERC721)
-   * @param epochManage Address of the epochManage contract
-   */
+     *
+     * @param metaDefender Address of the metaDefender contract
+     * @param liquidityCertificate Address of the liquidityCertificate contract(ERC721)
+     * @param policy Address of the policy contract(ERC721)
+     * @param epochManage Address of the epochManage contract
+     */
     function addMarket(
         address metaDefender,
         address liquidityCertificate,
         address policy,
         address epochManage
     ) external onlyOwner {
-        require(insuranceMarkets.add(metaDefender), "market already present");
+        require(insuranceMarkets.add(metaDefender), 'market already present');
         insuranceMarketsAddresses[metaDefender] = MarketAddresses(
             liquidityCertificate,
             policy,
@@ -50,11 +55,11 @@ contract MetaDefenderMarketsRegistry is Ownable, IMetaDefenderMarketsRegistry{
 
     /**
      * @dev Method to remove a market
-   *
-   * @param metaDefender Address of the metaDefender contract
-   */
+     *
+     * @param metaDefender Address of the metaDefender contract
+     */
     function removeMarket(address metaDefender) external onlyOwner {
-        require(insuranceMarkets.remove(metaDefender), "market not present");
+        require(insuranceMarkets.remove(metaDefender), 'market not present');
         delete insuranceMarketsAddresses[metaDefender];
 
         emit MarketRemoved(metaDefender);
@@ -62,12 +67,17 @@ contract MetaDefenderMarketsRegistry is Ownable, IMetaDefenderMarketsRegistry{
 
     /**
      * @dev Gets the list of addresses of deployments MetaDefender contracts
-   *
-   * @return Array of MetaDefender addresses
-   */
-    function getInsuranceMarkets() external view override returns (address[] memory) {
+     *
+     * @return Array of MetaDefender addresses
+     */
+    function getInsuranceMarkets()
+        external
+        view
+        override
+        returns (address[] memory)
+    {
         address[] memory list = new address[](insuranceMarkets.length());
-        for (uint i = 0; i < insuranceMarkets.length(); i++) {
+        for (uint256 i = 0; i < insuranceMarkets.length(); i++) {
             list[i] = insuranceMarkets.at(i);
         }
         return list;
@@ -75,19 +85,20 @@ contract MetaDefenderMarketsRegistry is Ownable, IMetaDefenderMarketsRegistry{
 
     /**
      * @dev Gets the addresses of the contracts associated to an InsuranceMarket contract
-   *
-   * @param insuranceMarketList Array of metaDefender contract addresses
-   * @return Array of struct containing the associated contract addresses
-   */
-    function getInsuranceMarketsAddresses(address[] calldata insuranceMarketList)
-    external
-    view
-    override
-    returns (MarketAddresses[] memory)
-    {
-        MarketAddresses[] memory marketAddresses = new MarketAddresses[](insuranceMarketList.length);
-        for (uint i = 0; i < insuranceMarketList.length; i++) {
-            marketAddresses[i] = insuranceMarketsAddresses[insuranceMarketList[i]];
+     *
+     * @param insuranceMarketList Array of metaDefender contract addresses
+     * @return Array of struct containing the associated contract addresses
+     */
+    function getInsuranceMarketsAddresses(
+        address[] calldata insuranceMarketList
+    ) external view override returns (MarketAddresses[] memory) {
+        MarketAddresses[] memory marketAddresses = new MarketAddresses[](
+            insuranceMarketList.length
+        );
+        for (uint256 i = 0; i < insuranceMarketList.length; i++) {
+            marketAddresses[i] = insuranceMarketsAddresses[
+                insuranceMarketList[i]
+            ];
         }
         return marketAddresses;
     }
